fix ddl

+++ derrick-migrate/src/migrate/pg.rs
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS {}(
   version bigint PRIMARY KEY,
   description text NOT NULL,
   content text NOT NULL,
-  duration_ms double precision NOT NULL,
+  duration_ms bigint NOT NULL,
   applied_at timestamptz NOT NULL DEFAULT now()
 );
 "#,
